Add SyncFusion Map
