+++ DisabilityServiceTools/Client/CampaignsWithoutNumber.Client.csproj
@@ -20,6 +20,7 @@
     <PackageReference Include="Microsoft.AspNetCore.SignalR.Client" Version="6.0.8" />
     <PackageReference Include="Microsoft.AspNetCore.WebUtilities" Version="2.2.0" />
     <PackageReference Include="MudBlazor" Version="6.0.14" />
+    <PackageReference Include="Syncfusion.Blazor.Maps" Version="20.2.0.50" />
     <PackageReference Include="System.Net.Http.Formatting.Extension" Version="5.2.3" />
     <PackageReference Include="System.Net.Http.Json" Version="6.0.0" />
   </ItemGroup>
