+++ DisabilityServiceTools/Client/Pages/WorldMap.razor
@@ -1,0 +1,12 @@
+﻿@page "/worldmap"
+@using Syncfusion.Blazor.Maps
+
+<h3>WorldMap</h3>
+
+<SfMaps>
+    
+</SfMaps>
+
+@code {
+
+}
