+++ DisabilityServiceTools/Client/Shared/NavMenu.razor
@@ -3,5 +3,6 @@
     <MudNavGroup Title="Characters" Expanded="true">
         <MudNavLink Href="/characterlist"  Match="NavLinkMatch.Prefix">Characters</MudNavLink>
         <MudNavLink Href="/characterclasslist"  Match="NavLinkMatch.Prefix">Character Classes</MudNavLink>
+        <MudNavLink Href="/worldmap"  Match="NavLinkMatch.Prefix">World Map</MudNavLink>
     </MudNavGroup>
-</MudNavMenu>
+</MudNavMenu>
